register vgg preset
