+++ keras_hub/api/models/__init__.py
@@ -299,7 +299,7 @@ from keras_hub.src.models.text_classifier_preprocessor import (
 from keras_hub.src.models.text_to_image import TextToImage
 from keras_hub.src.models.vgg.vgg_backbone import VGGBackbone
 from keras_hub.src.models.vgg.vgg_image_classifier import VGGImageClassifier
-from keras_hub.src.models.vgg.vgg_image_classifier import (
+from keras_hub.src.models.vgg.vgg_image_classifier_preprocessor import (
     VGGImageClassifierPreprocessor,
 )
 from keras_hub.src.models.vit_det.vit_det_backbone import ViTDetBackbone
