+++ keras_hub/src/models/vgg/vgg_image_classifier_test.py
@@ -3,6 +3,10 @@ import pytest
 
 from keras_hub.src.models.vgg.vgg_backbone import VGGBackbone
 from keras_hub.src.models.vgg.vgg_image_classifier import VGGImageClassifier
+from keras_hub.src.models.vgg.vgg_image_converter import VGGImageConverter
+from keras_hub.src.models.vgg.vgg_image_classifier_preprocessor import (
+    VGGImageClassifierPreprocessor,
+)
 from keras_hub.src.tests.test_case import TestCase
 
 
@@ -16,11 +20,16 @@ class VGGImageClassifierTest(TestCase):
             stackwise_num_filters=[2, 16, 16],
             image_shape=(8, 8, 3),
         )
+        image_converter = VGGImageConverter(image_size=(8, 8))
+        self.preprocessor = VGGImageClassifierPreprocessor(
+            image_converter=image_converter,
+        )
         self.init_kwargs = {
             "backbone": self.backbone,
             "num_classes": 2,
             "activation": "softmax",
             "pooling": "flatten",
+            "preprocessor": self.preprocessor,
         }
         self.train_data = (
             self.images,
@@ -28,9 +37,6 @@ class VGGImageClassifierTest(TestCase):
         )
 
     def test_classifier_basics(self):
-        pytest.skip(
-            reason="TODO: enable after preprocessor flow is figured out"
-        )
         self.run_task_test(
             cls=VGGImageClassifier,
             init_kwargs=self.init_kwargs,
@@ -44,4 +50,16 @@ class VGGImageClassifierTest(TestCase):
             cls=VGGImageClassifier,
             init_kwargs=self.init_kwargs,
             input_data=self.images,
+        )
+
+    # @pytest.mark.large
+    def test_smallest_preset(self):
+        # Test that our forward pass is stable!
+        image_batch = self.load_test_image()[None, ...] / 255.0
+        self.run_preset_test(
+            cls=VGGImageClassifier,
+            preset="vgg_11_imagenet",
+            input_data=image_batch,
+            expected_output_shape=(1, 1000),
+            expected_labels=[85],
         )
